Step by step tutorial Update examples and links
