--- docs/step_by_step.rst
@@ -1,32 +1,0 @@
-.. _examples:
-
-Step-by-step tutorial
-#####################
-
-The `samples directory </https://github.com/crowdresearch/daemo-api-client/tree/master/samples/>`_ in
-the Daemo API Client contains many examples to get you started, but here's
-a full tutorial that gives you an overview of the features.
-
-We are going to create, execute and monitor an image classification task.
-
-
-Installation
-------------
-
-Daemo API Client is tested on Mac OS X and Linux with the latest Python 2.7.
-To run a daemo client, you will need **python** and **virtualenv**.
-
-On Debian-based systems::
-
-    $ sudo apt-get install python-dev python-virtualenv
-
-Create a virtualenv and install *daemo-api-client* in it ::
-
-    $ virtualenv daemo
-    $ pip install daemo-api-client
-
-
-Usage
------
-
-
